requirements.txt
amqp==5.0.5
anyjson==0.3.3
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
Django==3.1.6
django-filter==2.4.0
djangorestframework==3.12.2
dnspython==1.16.0
eventlet==0.30.2
greenlet==1.0.0
httpie==2.4.0
idna==2.10
kombu==5.0.2
mysqlclient==2.0.3
prompt-toolkit==3.0.17
Pygments==2.8.1
PySocks==1.7.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
